% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_files}
\alias{concatenate_files}
\title{Concatenate the Contents of a Set of Files}
\usage{
concatenate_files(sources, target, headers = sprintf("# \%s ----",
  basename(sources)), preamble = NULL, postamble = NULL,
  verbose = configure_verbose())
}
\arguments{
\item{sources}{An \R list of files}

\item{target}{The file to use for generation.}

\item{headers}{Headers to be used for each file copied.}

\item{preamble}{Text to be included at the beginning of the document.}

\item{postamble}{Text to be included at the end of the document.}

\item{verbose}{Boolean; inform the user when the requested file is created?}
}
\description{
Given a set of files, concatenate their contents into
a single file.
}
