% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure-database.R
\docType{data}
\name{configure_database}
\alias{configure_database}
\alias{db}
\title{Retrieve the Global Configuration Database}
\format{An object of class \code{configure_database} of length 0.}
\usage{
configure_database()

db
}
\description{
Retrieve the global configuration database.
\code{db} is a helper alias for the database
returned by \code{configure_database()}.
}
\keyword{datasets}
