% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_directory}
\alias{configure_directory}
\title{Configure Files in a Directory}
\usage{
configure_directory(path = ".", config = configure_database(),
  verbose = configure_verbose())
}
\arguments{
\item{path}{The path to a directory in which files should be configured.}

\item{config}{The configuration database to be used.}

\item{verbose}{Boolean; report files as they are configured?}
}
\description{
This companion function to \code{\link[=configure_file]{configure_file()}} can be used to
configure all \code{.in} files within a directory.
}
\seealso{
Other configure: \code{\link{configure_file}}
}
