% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_file}
\alias{configure_file}
\title{Configure a File}
\usage{
configure_file(source, target = sub("[.]in$", "", source), config = NULL,
  verbose = getOption("configure.verbose", TRUE))
}
\arguments{
\item{source}{The file to be configured.}

\item{target}{The file to be generated.}

\item{config}{The configuration environment.}

\item{verbose}{Boolean; report files as they are configured?}
}
\description{
Configure a file, replacing any instances of `@`-delimited variables, e.g.
`@VAR@`, with the value of the variable called `VAR` in the associated
`config` environment.
}
