% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package}
\alias{package}
\alias{package-package}
\title{Configure R Packages for Installation}
\description{
Configure R packages for installation. This package provides a simple
framework, allowing one to prepare an R package for installation on
a particular platform using \R code.
}
