% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_cmd_config}
\alias{r_cmd_config}
\title{Execute R CMD config}
\usage{
r_cmd_config(...)
}
\arguments{
\item{...}{The names of potential configuration values.}
}
\description{
Read information about how \R is configured as through \code{R CMD config}.
}
