% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_config}
\alias{read_config}
\title{Read R Configuration for a Package}
\usage{
read_config(package = ".")
}
\arguments{
\item{package}{The path to an \R package's sources.}
}
\description{
Read the \R configuration, as through `R CMD config --all`.
}
