% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_r_config}
\alias{read_r_config}
\title{Read R Configuration for a Package}
\usage{
read_r_config(
  ...,
  package = Sys.getenv("R_PACKAGE_DIR", unset = "."),
  envir = configure_database(),
  verbose = configure_verbose()
)
}
\arguments{
\item{...}{The \R configuration values to read (as a character vector).
If empty, all values are read as through \verb{R CMD config --all}).}

\item{package}{The path to the \R package's sources.}

\item{envir}{The environment in which the configuration information should
be assigned. By default, the \code{\link[=configure_database]{configure_database()}} is populated with the
requested values.}

\item{verbose}{Boolean; notify the user as \R configuration is read?}
}
\description{
Read the \R configuration, as through \verb{R CMD config}.
}
