% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unity_build}
\alias{unity_build}
\title{Request a Unity Build}
\usage{
unity_build(use = TRUE)
}
\arguments{
\item{use}{Boolean; use a unity build?}
}
\description{
Perform a 'unity' build. All C / C++ source files will be concatenated into a
single file, and hence compiled as a single translation unit. This can
greatly increase compile times when compiling code that makes heavy use
of large header-only libraries (e.g. Rcpp or Boost).
}
