% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_configure}
\alias{use_configure}
\title{Add Configure Infrastructure to an R Package}
\usage{
use_configure(package = ".")
}
\arguments{
\item{package}{The path to the top-level directory of an \R package.}
}
\description{
Add the infrastructure needed to configure an R package.
}
