% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure-database.R
\name{configure_database}
\alias{configure_database}
\title{Retrieve the Global Configuration Database}
\usage{
configure_database()
}
\description{
Retrieve the global configuration database (as an \R environment).
}
\seealso{
Other configure-db: \code{\link{configure_define}}
}
