% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_file}
\alias{configure_file}
\title{Configure a File}
\usage{
configure_file(source, target = sub("[.]in$", "", source),
  config = configure_database(), lhs = "@", rhs = "@",
  verbose = configure_verbose())
}
\arguments{
\item{source}{The file to be configured.}

\item{target}{The file to be generated.}

\item{config}{The configuration database.}

\item{lhs}{The left-hand side marker; defaults to \code{@}.}

\item{rhs}{The right-hand side marker; defaults to \code{@}.}

\item{verbose}{Boolean; report files as they are configured?}
}
\description{
Configure a file, replacing (by default) any instances of \code{@}-delimited
variables, e.g. \code{@VAR@}, with the value of the variable called \code{VAR} in the
associated \code{config} environment.
}
\seealso{
Other configure: \code{\link{configure_directory}}
}
\concept{configure}
